+++ scripts/torch_cli/cli/lib/vllm/setup_vllm.sh
