+++ scripts/torch_cli/cli/test/__init__.py
@@ -1,0 +1,6 @@
+
+
+from cli.test.external.controller import ExternalTestController
+from cli.test.main import TestController
+
+TEST_CONTROLLERS = [TestController,ExternalTestController]
